New low-level API not based on iou.
This makes setting the user_data field private to ringbahn, so it
is safe to complete a CQE assuming it contains a Completion. With
some other changes around ring ownership, it makes the demo driver 100%
safe code.

The full iou API has not yet been implemented, just enough to implement
the demo driver.

The transition has begun to try to allow events to request multiple SQEs
from the SQ, allowing them to create linked events in a single go. The
first multi-SQE API will be the `hard_linked` API, which allows multiple
events to be linked together with `IORING_SQE_HARDLINK`. The results of
all but the final event will be ignored; the completion will only awaken
on the final event. This functionality is currently not correct
implemented.

+++ src/submission.rs
@@ -40,7 +40,7 @@ impl<E, D> Future for Submission<E, D> where
     E: Event,
     D: Drive,
 {
-    type Output = (E, io::Result<usize>);
+    type Output = (E, io::Result<u32>);
 
     fn poll(self: Pin<&mut Self>, ctx: &mut Context<'_>) -> Poll<Self::Output> {
         let (ring, event) = self.split();
